requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
av==13.1.0
bitarray==3.0.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Cython==3.0.11
decorator==5.1.1
edge-tts==6.1.18
exceptiongroup==1.2.2
fairseq==0.12.2
faiss-cpu==1.7.3
fastapi==0.115.5
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
h11==0.14.0
hydra-core==1.0.7
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
loguru==0.7.2
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numpy==1.23.5
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.0.6
packaging==24.2
platformdirs==4.3.6
pooch==1.8.2
portalocker==2.10.1
praat-parselmouth==0.4.5
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.0.0
python-ffmpeg==2.0.12
python-multipart==0.0.17
pyworld==0.3.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rvc-python==0.1.5
sacrebleu==2.4.3
scikit-learn==1.5.2
scipy==1.14.1
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
starlette==0.41.2
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
torch==2.5.1
torchaudio==2.5.1
torchcrepe==0.0.23
tqdm==4.67.0
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
yarl==1.17.1
